editorconfig and crlf

+++ src/Unnamed/components/Transform.h
@@ -153,17 +153,17 @@ private:
         // roll (x-axis rotation)
         double sinr_cosp = 2 * (quaternion.w * quaternion.x + quaternion.y * quaternion.z);
         double cosr_cosp = 1 - 2 * (quaternion.x * quaternion.x + quaternion.y * quaternion.y);
-        angles.x = std::atan2(sinr_cosp, cosr_cosp);
+        angles.x = static_cast<float>(std::atan2(sinr_cosp, cosr_cosp));
 
         // pitch (y-axis rotation)
         double sinp = std::sqrt(1 + 2 * (quaternion.w * quaternion.y - quaternion.x * quaternion.z));
         double cosp = std::sqrt(1 - 2 * (quaternion.w * quaternion.y - quaternion.x * quaternion.z));
-        angles.y = 2 * std::atan2(sinp, cosp) - XM_PI / 2;
+        angles.y = 2.0f * static_cast<float>(std::atan2(sinp, cosp)) - XM_PI / 2;
 
         // yaw (z-axis rotation)
         double siny_cosp = 2 * (quaternion.w * quaternion.z + quaternion.x * quaternion.y);
         double cosy_cosp = 1 - 2 * (quaternion.y * quaternion.y + quaternion.z * quaternion.z);
-        angles.z = std::atan2(siny_cosp, cosy_cosp);
+        angles.z = static_cast<float>(std::atan2(siny_cosp, cosy_cosp));
 
         return angles;
     }
